requirements.txt
fastapi
uvicorn
sqlalchemy
google-generativeai
streamlit
python-dotenv
groq
